Made code more readable, one bug fix
Changed methods which that weren't intuitive before to be more
intuitive. For example an overloaded Move method has been created for
the Character class such that you only need row and column to use it.
- Added method overloads to methods that required a layer, row and
column. Changed them such that row and column are the only ones
required.  This is done making the assumption that those methods will
only occur on a single layer.
- Added parameter mainLayer to TileMap. This is the layer where
characters can move.
- Fixed bug when Character 1 moved over Character 2 in its path caused
Character 2's tile to be "empty" even when Character 2 was still there.